+++ SmtpToRest.Docker/Program.cs
@@ -1,7 +1,10 @@
 using Microsoft.AspNetCore.Builder;
 using Microsoft.Extensions.DependencyInjection;
-using SmtpToRest;
+using SmtpToRest.Config;
 using SmtpToRest.Docker;
+using SmtpToRest.Processing;
+using SmtpToRest.Rest;
+using SmtpToRest.Services.Smtp;
 
 var builder = WebApplication.CreateBuilder(args);
 
@@ -10,7 +13,8 @@ builder.Services
     .AddSingleton<IConfigurationFileReader, DockerConfigurationFileReader>()
     .AddSingleton<IMessageStoreFactory, DefaultMessageStoreFactory>()
     .AddSingleton<ISmtpServerFactory, DefaultSmtpServerFactory>()
-    .AddSingleton<IRestClient, RestClient>()
+    .AddSingleton<IMessageProcessor, DefaultMessageProcessor>()
+	.AddSingleton<IRestClient, RestClient>()
     .AddSingleton<IHttpClientFactory, DefaultHttpClientFactory>()
     .AddHostedService<SmtpServerBackgroundService>();
