+++ SmtpToRest.WindowsService/Program.cs
@@ -1,5 +1,9 @@
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
+using SmtpToRest.Config;
+using SmtpToRest.Processing;
+using SmtpToRest.Rest;
+using SmtpToRest.Services.Smtp;
 
 namespace SmtpToRest.WindowsService;
 
@@ -12,13 +16,14 @@ public class Program
 
     public static IHostBuilder CreateHostBuilder(string[] args) =>
         Host.CreateDefaultBuilder(args)
-            .ConfigureServices((hostContext, services) =>
+            .ConfigureServices((_, services) =>
             {
                 services
                     .AddSingleton<IConfiguration, Configuration>()
                     .AddSingleton<IConfigurationFileReader, DefaultConfigurationFileReader>()
                     .AddSingleton<IMessageStoreFactory, DefaultMessageStoreFactory>()
                     .AddSingleton<ISmtpServerFactory, DefaultSmtpServerFactory>()
+                    .AddSingleton<IMessageProcessor, DefaultMessageProcessor>()
                     .AddSingleton<IRestClient, RestClient>()
                     .AddSingleton<IHttpClientFactory, DefaultHttpClientFactory>()
                     .AddHostedService<SmtpServerBackgroundService>();
