+++ SmtpToRest/Config/IConfigurationFileReader.cs
@@ -1,0 +1,6 @@
+﻿namespace SmtpToRest.Config;
+
+public interface IConfigurationFileReader
+{
+    string Read();
+}
