--- SmtpToRest/IConfigurationFileReader.cs
@@ -1,6 +1,0 @@
-﻿namespace SmtpToRest;
-
-public interface IConfigurationFileReader
-{
-    string Read(string path);
-}
