+++ SmtpToRest/Rest/HttpRequestHeadersAdapter.cs
@@ -1,6 +1,6 @@
 ﻿using System.Net.Http.Headers;
 
-namespace SmtpToRest;
+namespace SmtpToRest.Rest;
 
 public class HttpRequestHeadersAdapter : IHttpRequestHeaders
 {
@@ -9,7 +9,7 @@ public class HttpRequestHeadersAdapter : IHttpRequestHeaders
         get => _adaptee.Authorization;
         set => _adaptee.Authorization = value;
     }
-        
+
     private readonly HttpRequestHeaders _adaptee;
 
     public HttpRequestHeadersAdapter(HttpRequestHeaders adaptee)
