+++ SmtpToRest/Rest/IHttpClientFactory.cs
@@ -1,4 +1,4 @@
-﻿namespace SmtpToRest;
+﻿namespace SmtpToRest.Rest;
 
 public interface IHttpClientFactory
 {
