+++ SmtpToRest/Rest/IRestClient.cs
@@ -1,8 +1,9 @@
 ﻿using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
+using SmtpToRest.Config;
 
-namespace SmtpToRest;
+namespace SmtpToRest.Rest;
 
 public interface IRestClient
 {
