+++ SmtpToRest/Services/Smtp/DefaultSmtpServerFactory.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using SmtpServer;
 
-namespace SmtpToRest;
+namespace SmtpToRest.Services.Smtp;
 
 public class DefaultSmtpServerFactory : ISmtpServerFactory
 {
